Consolidated PureResourcePath into ResourcePath.
