+++ st3/sublime_lib/__init__.py
@@ -4,5 +4,4 @@ from .view_stream import ViewStream  # noqa: F401
 from .output_panel import OutputPanel  # noqa: F401
 from .show_selection_panel import show_selection_panel  # noqa: F401
 from .window_utils import new_window, close_window  # noqa: F401
-from .pure_resource_path import PureResourcePath  # noqa: F401
 from .resource_path import ResourcePath  # noqa: F401
